+++ GeerWeekApp/main/java/ca/ippetad/geerweekapp/view/EventsFragment.java
@@ -1,4 +1,4 @@
-package ca.ippetad.geerweekapp.ca.ippetad.geerweekapp.view;
+package ca.ippetad.geerweekapp.view;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
@@ -10,14 +10,14 @@ import ca.ippetad.geerweekapp.R;
 /**
  * Created by James on 2/8/14.
  */
-public class MiscFragment  extends Fragment {
+public class EventsFragment extends Fragment {
 
-    public MiscFragment() {}
+    public EventsFragment() {}
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        View rootView = inflater.inflate(R.layout.fragment_misc, container, false);
+        View rootView = inflater.inflate(R.layout.fragment_events, container, false);
 
 
         return rootView;
